Refactor Project Structure
Moved chat.py, synthesize.py, and  transcribe.py into modules folder. And updated main.py imports.
