+++ main.py
@@ -1,7 +1,7 @@
 import argparse
-from transcribe import transcribe_file
-from chat import get_completion
-from synthesize import synthesize_speech
+from modules.transcribe import transcribe_file
+from modules.chat import get_completion
+from modules.synthesize import synthesize_speech
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="音声ファイルを文字起こしして翻訳し、音声に合成するスクリプト")
